requirements.txt
sqlalchemy == 2.0.35
pydantic == 2.9.2
colorama == 0.4.6
python-jose ==3.3.0
passlib==1.7.4
bcrypt==4.2.0



# FastAPI is a modern, fast (high-performance), web framework for building APIs with Python 3.6+ based on standard Python type hints.
anyio ~= 4.4.0
click ~= 8.1.7
exceptiongroup ~= 1.2.2
fastapi == 0.115.5
h11 ~=0.14.0
httptools ~=0.6.1
idna ~=3.10
python-dotenv ~= 1.0.1
pyyaml ~=6.0.2
sniffio ~=1.3.1
starlette == 0.41.2
uvicorn == 0.32.0
watchfiles ~=0.24.0
websockets == 14.1
python-multipart ~=0.0.9

requests == 2.32.3
openpyxl == 3.1.5
#remove
aiohttp == 3.11.7
aiomysql == 0.2.0